+++ test-parts/test-graph-parse.js
@@ -6,6 +6,8 @@ const parseGraph = require("../src/parse-graph");
 
 const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 const maxEdges = 300;
+const formatErrorText = "Could not parse input into a valid graph.";
+const arrayErrorText = "Parsed graph must have multiple nodes and edges.";
 
 var parseOutputData = null;
 
@@ -17,6 +19,7 @@ function runTests()
 		loadOutputData();
 		handleValidCase();
 		handleIgnoreData();
+		handleInvalidCases();
 	});
 }
 
@@ -138,6 +141,80 @@ function handleIgnoreData()
 }
 
 
+function handleInvalidCases()
+{
+	describe("Invalid Cases", function()
+	{
+		it("Invalid Type", function()
+		{
+			callInvalidEntry(null, "Cannot read property 'replace' of null");
+		});
+		
+		it("Empty String", function()
+		{
+			callInvalidEntry("", formatErrorText);
+		});
+		
+		it("Wrong String Format", function()
+		{
+			callInvalidEntry("Not Graph: XYZ", formatErrorText);
+		});
+		
+		it("Single Edge Entry", function()
+		{
+			callInvalidEntry("AB1", formatErrorText);
+		});
+		
+		it("Single Edge Output", function()
+		{
+			callInvalidEntry("AB1, AB2, AB3", arrayErrorText);
+		});
+		
+		it("Single Node", function()
+		{
+			callInvalidEntry("AA1, AA2, AA3", arrayErrorText);
+		});
+		
+		it("Single Node And Edge", function()
+		{
+			callInvalidEntry("AA4", formatErrorText);
+		});
+		
+		it("Missing Destination", function()
+		{
+			callInvalidEntry("AB2, C3, DE4", formatErrorText);
+		});
+		
+		it("Missing Distance", function()
+		{
+			callInvalidEntry("AB1, CD, EF3", formatErrorText);
+		});
+		
+		it("Negative Distance", function()
+		{
+			callInvalidEntry("AB5, CD1, EF-5", formatErrorText);
+		});
+		
+		it("Decimal Distance", function()
+		{
+			callInvalidEntry("AB1, CD1.5, EF2", formatErrorText);
+		});
+		
+		it("Unsupported Node Character", function()
+		{
+			callInvalidEntry("QR2, $T3", formatErrorText);
+		});
+		
+		it("Unsupported Distance Character", function()
+		{
+			callInvalidEntry("UV8, TSx", formatErrorText);
+		});
+		
+		
+	});
+}
+
+
 function getNodeList(entryStr)
 {
 	var listRes = entryStr.split("");
@@ -198,6 +275,39 @@ function defineMaxEdgesGraph()
 }
 
 
+function callInvalidEntry(entryStr, desiredMessage)
+{
+	var parseSuccessful = false;
+	var correctError = false;
+	
+	try
+	{
+		parseGraph.performParsing(entryStr);
+		parseSuccessful = true;
+	}
+	catch(parseErr)
+	{
+		parseSuccessful = false;
+		correctError = parseErr.message.startsWith(desiredMessage);
+	}
+	
+	
+	if (parseSuccessful === true)
+	{
+		throw new Error("No error was thrown.");
+	}
+	else if (correctError === true)
+	{
+		expect(true).to.be.true;
+	}
+	else
+	{
+		flagIncorrectError(desiredMessage);
+	}
+	
+}
+
+
 
 function checkParseResult(parseObj)
 {
@@ -223,6 +333,13 @@ function checkGraphContents(parseObj, nodeArr, edgeArr)
 {
 	expect(parseObj.nodes).to.deep.equal(nodeArr);
 	expect(parseObj.edges).to.deep.equal(edgeArr);
+}
+
+
+function flagIncorrectError(vExp)
+{
+	var preparedText = ["Incorrect error thrown.\r\n", "Should had been: '", vExp, "'"].join("");
+	throw new Error(preparedText);
 }